+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# v2.0 (2021-Sep-30)
+* Added a mail ticket handler for DM messages
+* Added a DM scanning system for profile screenshots
+* Integrated the two, so that failed screenshots are passed on to the mail section
+* Added dmMail and dmScanning toggles to enable these features
+* Added a Pokenav message deletion filter
+* Added `]add` and `]remove` to add channels to the filter list
+* Added "DM" to some error messages
+* Made level 29 screenshots redirect to mail
+* Added YPOBEI to the rescan word list
+
+
+
 # v1.8 (2021-Sep-10)
 * Desynced and reworked **logImg** for a potential 100% speed increase
 * Added configurable message text in ./server/messagetxt.js